+++ Content.Shared/_WL/_Offbrand/StatusEffects/GibOnDamageStatusEffectSystem.cs
@@ -1,7 +1,5 @@
 using Content.Shared.Damage.Systems;
 using Content.Shared.Damage.Components;
-using Content.Shared.Damage.Prototypes;
-using Content.Shared.Damage;
 using Content.Shared.Popups;
 using Content.Shared.Gibbing;
 using Content.Shared.StatusEffectNew;
