+++ Content.Shared/_WL/_Offbrand/Surgery/SurgeryTargetComponent.cs
@@ -1,7 +1,3 @@
-using Content.Shared.Construction.Prototypes;
-using Robust.Shared.Prototypes;
-using Robust.Shared.Serialization;
-
 namespace Content.Shared._WL._Offbrand.Surgery;
 
 [RegisterComponent]
