+++ Content.Shared/_WL/_Offbrand/Wounds/WoundSpecifier.cs
@@ -1,7 +1,4 @@
 using Content.Shared.Damage;
-using Content.Shared.Damage.Prototypes;
-using Content.Shared.FixedPoint;
-using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization;
